sense-python/src/requirements.txt
RPi.bme280==0.2.4
adafruit-circuitpython-busdevice>=4.1.1

# Adafruit-GPIO==1.0.3
Adafruit-PlatformDetect>=2.1.0
Adafruit-PureIO>=1.0.4
Flask==2.2.2
gpiozero==1.6.2
grpcio==1.50.0
protobuf==4.21.9
RPi.GPIO==0.7.1

#;  platform_machine=='armv7l' or platform_machine=='armv6l'
smbus2==0.4.2
stringify==0.1.1
#picamera==1.13